Добавил отмену операции. Изменил вычисление прогресса. Добавил чуть больше эмпатии

+++ BlockArchiver/BlockArchiver/ProgressEventArgs.cs
@@ -1,8 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace BlockArchiver
 {
@@ -10,9 +6,9 @@ namespace BlockArchiver
     {
         public int Progress { get; set; }
 
-        public ProgressEventArgs(long current, long total)
+        public ProgressEventArgs(int current, int total)
         {
-            Progress = (int)(current * 100 / total);
+            Progress = current * 100 / total;
         }
     }
 }
